+++ src/main/resources/db/migration/V02__Remove_outbox.sql
@@ -2,3 +2,4 @@ DROP TABLE outbox;
 
 ALTER TABLE auth_event
 ADD COLUMN was_sent boolean NOT NULL DEFAULT FALSE;
+--todo should I rename this file or move editing auth_event to another migration?